Refactor: Carousel 스타일 폴더 추가 및 경로 수정
