+++ FE/sidedish/src/components/carouselLibR/Badge.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { BadgeBox } from "style/Badge";
+import { BadgeBox } from "style/carouselLibR/Badge";
 
 const Badge = ({ name }) => (
 	<BadgeBox color={name}>
