+++ FE/sidedish/src/components/carouselLibR/CarouselHeader.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { Header, Title, SubTitle } from "style/CarouselHeader";
+import { Header, Title, SubTitle } from "style/carouselLibR/CarouselHeader";
 
 const CarouselHeader = ({ title, subTitle }) => (
 	<Header>
